move ising gates to parametric
