+++ pennylane/ops/qubit/qchem_ops.py
@@ -36,177 +36,6 @@ a = np.pi / 2
 b = 3 * np.pi / 2
 four_term_grad_recipe = ([[c1, 1, a], [-c1, 1, -a], [-c2, 1, b], [c2, 1, -b]],)
 
-
-class IsingXX(Operation):
-    r"""IsingXX(phi, wires)
-    Ising XX coupling gate
-
-    .. math:: XX(\phi) = \begin{bmatrix}
-            \cos(\phi / 2) & 0 & 0 & -i \sin(\phi / 2) \\
-            0 & \cos(\phi / 2) & -i \sin(\phi / 2) & 0 \\
-            0 & -i \sin(\phi / 2) & \cos(\phi / 2) & 0 \\
-            -i \sin(\phi / 2) & 0 & 0 & \cos(\phi / 2)
-        \end{bmatrix}.
-
-    **Details:**
-
-    * Number of wires: 2
-    * Number of parameters: 1
-    * Gradient recipe: :math:`\frac{d}{d\phi}f(XX(\phi)) = \frac{1}{2}\left[f(XX(\phi +\pi/2)) - f(XX(\phi-\pi/2))\right]`
-      where :math:`f` is an expectation value depending on :math:`XX(\phi)`.
-
-    Args:
-        phi (float): the phase angle
-        wires (int): the subsystem the gate acts on
-    """
-    num_params = 1
-    num_wires = 2
-    par_domain = "R"
-    grad_method = "A"
-    generator = [
-        np.array([[0, 0, 0, 1], [0, 0, 1, 0], [0, 1, 0, 0], [1, 0, 0, 0]]),
-        -1 / 2,
-    ]
-
-    @classmethod
-    def _matrix(cls, *params):
-        phi = params[0]
-        c = math.cos(phi / 2)
-        s = math.sin(phi / 2)
-
-        return np.array(
-            [
-                [c, 0, 0, -1j * s],
-                [0, c, -1j * s, 0],
-                [0, -1j * s, c, 0],
-                [-1j * s, 0, 0, c],
-            ]
-        )
-
-    @staticmethod
-    def decomposition(phi, wires):
-        decomp_ops = [
-            CNOT(wires=wires),
-            RX(phi, wires=[wires[0]]),
-            CNOT(wires=wires),
-        ]
-        return decomp_ops
-
-    def adjoint(self):
-        (phi,) = self.parameters
-        return IsingXX(-phi, wires=self.wires)
-
-
-class IsingYY(Operation):
-    r"""IsingYY(phi, wires)
-    Ising YY coupling gate
-
-    .. math:: \mathtt{YY}(\phi) = \begin{bmatrix}
-        \cos(\phi / 2) & 0 & 0 & i \sin(\phi / 2) \\
-        0 & \cos(\phi / 2) & -i \sin(\phi / 2) & 0 \\
-        0 & -i \sin(\phi / 2) & \cos(\phi / 2) & 0 \\
-        i \sin(\phi / 2) & 0 & 0 & \cos(\phi / 2)
-        \end{bmatrix}.
-
-    **Details:**
-
-    * Number of wires: 2
-    * Number of parameters: 1
-    * Gradient recipe: :math:`\frac{d}{d\phi}f(YY(\phi)) = \frac{1}{2}\left[f(YY(\phi +\pi/2)) - f(YY(\phi-\pi/2))\right]`
-      where :math:`f` is an expectation value depending on :math:`YY(\phi)`.
-
-    Args:
-        phi (float): the phase angle
-        wires (int): the subsystem the gate acts on
-    """
-    num_params = 1
-    num_wires = 2
-    par_domain = "R"
-    grad_method = "A"
-    generator = [
-        np.array([[0, 0, 0, -1], [0, 0, 1, 0], [0, 1, 0, 0], [-1, 0, 0, 0]]),
-        -1 / 2,
-    ]
-
-    @staticmethod
-    def decomposition(phi, wires):
-        return [
-            qml.CY(wires=wires),
-            qml.RY(phi, wires=[wires[0]]),
-            qml.CY(wires=wires),
-        ]
-
-    @classmethod
-    def _matrix(cls, *params):
-        phi = params[0]
-        cos = np.cos(phi / 2)
-        isin = 1.0j * np.sin(phi / 2)
-        return np.array(
-            [
-                [cos, 0.0, 0.0, isin],
-                [0.0, cos, -isin, 0.0],
-                [0.0, -isin, cos, 0.0],
-                [isin, 0.0, 0.0, cos],
-            ],
-            dtype=complex,
-        )
-
-    def adjoint(self):
-        (phi,) = self.parameters
-        return IsingYY(-phi, wires=self.wires)
-
-
-class IsingZZ(Operation):
-    r""" IsingZZ(phi, wires)
-    Ising ZZ coupling gate
-
-    .. math:: ZZ(\phi) = \begin{bmatrix}
-        e^{-i \phi / 2} & 0 & 0 & 0 \\
-        0 & e^{i \phi / 2} & 0 & 0 \\
-        0 & 0 & e^{i \phi / 2} & 0 \\
-        0 & 0 & 0 & e^{-i \phi / 2}
-        \end{bmatrix}.
-
-    **Details:**
-
-    * Number of wires: 2
-    * Number of parameters: 1
-    * Gradient recipe: :math:`\frac{d}{d\phi}f(ZZ(\phi)) = \frac{1}{2}\left[f(ZZ(\phi +\pi/2)) - f(ZZ(\phi-\pi/2))\right]`
-      where :math:`f` is an expectation value depending on :math:`ZZ(\theta)`.
-
-    Args:
-        phi (float): the phase angle
-        wires (int): the subsystem the gate acts on
-    """
-    num_params = 1
-    num_wires = 2
-    par_domain = "R"
-    grad_method = "A"
-    generator = [
-        np.array([[1, 0, 0, 0], [0, -1, 0, 0], [0, 0, -1, 0], [0, 0, 0, 1]]),
-        -1 / 2,
-    ]
-
-    @staticmethod
-    def decomposition(phi, wires):
-        return [
-            qml.CNOT(wires=wires),
-            qml.RZ(phi, wires=[wires[1]]),
-            qml.CNOT(wires=wires),
-        ]
-
-    @classmethod
-    def _matrix(cls, *params):
-        phi = params[0]
-        pos_phase = np.exp(1.0j * phi / 2)
-        neg_phase = np.exp(-1.0j * phi / 2)
-        return np.diag([neg_phase, pos_phase, pos_phase, neg_phase])
-
-    def adjoint(self):
-        (phi,) = self.parameters
-        return IsingZZ(-phi, wires=self.wires)
-
-
 class SingleExcitation(Operation):
     r"""SingleExcitation(phi, wires)
     Single excitation rotation.
